requirements.txt
tqdm
requests

http://download.pytorch.org/whl/cu90/torch-0.3.1-cp36-cp36m-linux_x86_64.whl
torchvision

face_recognition
scikit-image
